fastapi_postgres_app/requirements.txt
asyncpg==0.27.0
fastapi==0.89.1
fastapi-sqlalchemy==0.2.1
psycopg2-binary==2.9.5
uvicorn==0.20.0